Create account page // Fixing generate invite link // Group participants dialog

+++ app/api/invitations/route.ts
@@ -14,6 +14,43 @@ export async function GET(request: NextRequest) {
   }
 
   try {
+    const res = await client.sql`
+      SELECT invite_code
+      FROM invitations
+      WHERE group_id=${group_id}
+    `;
+
+    const { invite_code } = res.rows[0] as InvitationResponse;
+
+    return NextResponse.json({
+      data: {
+        invite_code,
+        group_id,
+      },
+    });
+  } catch (err) {
+    console.log(err);
+    return NextResponse.json(
+      { error: "Something went wrong! Try again later." },
+      { status: 500 }
+    );
+  } finally {
+    client.release();
+  }
+}
+
+export async function POST(request: NextRequest) {
+  const client = await db.connect();
+  const group_id = request.cookies.get("group_id")?.value;
+
+  if (!group_id) {
+    return NextResponse.json(
+      { error: "Group ID not found or invalid." },
+      { status: 404 }
+    );
+  }
+
+  try {
     await client.sql`
       DELETE FROM invitations
       WHERE group_id = ${group_id}
